Lifted state up to the board component
Before, each square component managed itself, but how do each communicate with each other and tell the user who is the winner? This is when we must "lift the state up". The nearest parent component of all squares is the board component. We will manage state from the board component and pass down values and event handling function as props. When a square is clicked, it will invoke the function inside the board component and this will then update itself and pass down new props.